+++ ibllib/qc/task_qc_viewer/__init__.py
@@ -1,0 +1,1 @@
+"""Interactive task QC viewer."""